fix: 에러핸들링 보완

+++ server/src/services/user.service.ts
@@ -16,7 +16,7 @@ export const createUser = async (userData: Partial<User>) => {
   newUser.accessLevel = 1;
 
   await userRepository.save(newUser);
-  return { message: 'User created successfully' };
+  return { message: '회원 가입이 완료되었습니다.' };
 };
 
 // * 로그인
@@ -24,13 +24,13 @@ export const login = async (email: string, password: string) => {
   const user = await userRepository.findOne({ where: { email } });
 
   if (!user) {
-    throw new HttpError(404, 'User not found');
+    throw new HttpError(404, '사용자를 찾을 수 없습니다.');
   }
 
   const isPasswordValid = await bcrypt.compare(password, user.password);
 
   if (!isPasswordValid) {
-    throw new HttpError(400, 'Email or password is incorrect');
+    throw new HttpError(400, '이메일 또는 비밀번호가 올바르지 않습니다.');
   }
 
   const token = jwt.sign(
@@ -53,7 +53,7 @@ export const getMyInfo = async (userId: number) => {
   const user = await userRepository.findOne({ where: { id: userId }, select: ['id', 'name', 'email', 'status'] });
 
   if (!user) {
-    throw new HttpError(404, 'User not found');
+    throw new HttpError(404, '사용자를 찾을 수 없습니다.');
   }
   return user;
 };
@@ -63,7 +63,7 @@ export const updateMyInfo = async (userId: number, updateData: Partial<User>) =>
   const user = await userRepository.findOne({ where: { id: userId } });
 
   if (!user) {
-    throw new HttpError(404, 'User not found');
+    throw new HttpError(404, '사용자를 찾을 수 없습니다.');
   }
 
   if (updateData.password) {
@@ -76,5 +76,11 @@ export const updateMyInfo = async (userId: number, updateData: Partial<User>) =>
 
 // * 유저 삭제
 export const deleteUser = async (userId: number) => {
+  const user = await userRepository.findOne({ where: { id: userId } });
+
+  if (!user) {
+    throw new HttpError(404, '사용자를 찾을 수 없습니다.');
+  }
+
   await userRepository.delete(userId);
 };
